#1875 - Logout from external authentication
- Move preauth settings to security module
- Change prefix of logout url setting to "auth.preauth"
- Updated documentation accordingly
