+++ inception/inception-security/src/main/java/de/tudarmstadt/ukp/clarin/webanno/security/config/PreauthenticationPropertiesImpl.java
@@ -1,0 +1,49 @@
+/*
+ * Licensed to the Technische Universität Darmstadt under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The Technische Universität Darmstadt
+ * licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.tudarmstadt.ukp.clarin.webanno.security.config;
+
+import java.util.Optional;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import de.tudarmstadt.ukp.clarin.webanno.security.preauth.ShibbolethRequestHeaderAuthenticationFilter;
+
+/**
+ * This class is exposed as a Spring Component via {@link SecurityAutoConfiguration}.
+ * <p>
+ * There are additional pre-auth settings which are handled directly by the
+ * {@link ShibbolethRequestHeaderAuthenticationFilter}.
+ */
+@ConfigurationProperties("auth.preauth")
+public class PreauthenticationPropertiesImpl
+    implements PreauthenticationProperties
+{
+    private String logoutUrl;
+
+    @Override
+    public Optional<String> getLogoutUrl()
+    {
+        return Optional.ofNullable(logoutUrl);
+    }
+
+    @Override
+    public void setLogoutUrl(String aLogoutUrl)
+    {
+        logoutUrl = aLogoutUrl;
+    }
+}
