+++ inception/inception-security/src/main/java/de/tudarmstadt/ukp/clarin/webanno/security/config/SecurityAutoConfiguration.java
@@ -39,7 +39,8 @@ import de.tudarmstadt.ukp.clarin.webanno.security.UserDao;
 import de.tudarmstadt.ukp.clarin.webanno.security.UserDaoImpl;
 
 @Configuration
-@EnableConfigurationProperties(SecurityPropertiesImpl.class)
+@EnableConfigurationProperties({ SecurityPropertiesImpl.class,
+        PreauthenticationPropertiesImpl.class })
 public class SecurityAutoConfiguration
 {
     private @PersistenceContext EntityManager entityManager;
