+++ inception/inception-security/src/main/resources/META-INF/asciidoc/admin-guide/security-preauth.adoc
@@ -73,15 +73,15 @@ NOTE: The roles added through `auth.user.<username>.roles` properties are *not* 
 | _<none>_
 | ROLE_PROJECT_CREATOR
 
+| auth.preauth.logoutUrl
+| URL to call after logging out to also sign out of external authentication
+| _<none>_
+| https://your-idp.com/Shibboleth.sso/Logout
+
 | auth.user._<username>_.roles
 | Extra roles for user (comma separated)
 | _<none>_
 | ROLE_ADMIN
-
-| auth.logoutUrl
-| URL to call after logging out
-| _<none>_
-| https://your-idp.com/Shibboleth.sso/Logout
 
 |===
